구매내역에서 경매진행중(미완)과 결제완료로 구분함

+++ src/main/java/project/spring/nft/persistence/PaymentDAO.java
@@ -17,5 +17,5 @@ public interface PaymentDAO {
 	List<PaymentVO> selectAll(String memberId);
 	
 	//판매수익
-	int selectProfit(String memberId);
+	Double selectProfit(String memberId);
 }
